# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Syncitt
# during dev, should point to your local machine
host: localhost:5095
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/dashboardInspection:
    # binds a127 app logic to a route
    x-swagger-router-controller: dashboard
    post:
      tags: ["Dashboard"]
      description: Inspection list on dashboard
      summary: Inspection calender data on dashboard
      # used as the method name of the controller
      operationId: dashboardInspection
      parameters:
        - name: body
          in: body
          description: request_by_role => Active role id, request_by_id => Logged in user id and agency_id => looged in user agency id
          schema:
            $ref: "#/definitions/dashboardInspectionDefination"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getStatisticsData:
    # binds a127 app logic to a route
    x-swagger-router-controller: dashboard
    post:
      tags: ["Dashboard"]
      description: List of statistics data on dahsboard
      summary: List of statistics data on dahsboard
      # used as the method name of the controller
      operationId: getStatisticsData
      parameters:
        - name: body
          in: body
          description: Agency Id
          schema:
            $ref: "#/definitions/getStatisticsDataDefination"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/contact_us:
    # binds a127 app logic to a route
    x-swagger-router-controller: contactUs
    post:
      tags: ["contact us"]
      description: To send email to admin for contact us
      summary: To send email to admin for contact us
      # used as the method name of the controller
      operationId: contactUs
      parameters:
        - name: body
          in: body
          description: contact us
          schema:
            $ref: "#/definitions/contact_us"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  contact_us:
    type: object
    properties:
      firstname:
        type: string
      lastname:
        type: string
      email:
        type: string
      role:
        type: string
      mobile_no:
        type: string
      comment:
        type: string
      agency_name:
        type: string
...
  getStatisticsDataDefination:
    type: object
    properties:
      agency_id:
        type: string
      request_by_role:
        type: string
      user_id:
        type: string
  # getAdminStatisticsData:
  #   type: object
  #   properties:
  #     _id:
  #       type: string
...
  dashboardInspectionDefination:
    type: object
    properties:
      request_by_id:
        type: string
      request_by_role:
        type: string
      agency_id:
        type: string
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string